add basic svelte editor component and use it for preview
