+++ frontend/src/documents/DocumentPreview.svelte
@@ -1,4 +1,6 @@
 <script>
+  import Editor from "../editor/Editor.svelte";
+  import { fetch, handleText } from "../lib/fetch";
   import { ext } from "../lib/paths";
 
   import { baseURL } from "../stores/url";
@@ -19,6 +21,12 @@
 
 {#if extension === 'pdf'}
   <object title={filename} data={url} />
+{:else if ['csv', 'txt'].includes(extension)}
+  {#await fetch(url).then(handleText)}
+    Loading...
+  {:then value}
+    <Editor {value} />
+  {/await}
 {:else}
   Preview for file `{filename}` with file type `{extension}` is not implemented
 {/if}
